// backend/src/generated-api/models/create-season-request-body.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Survivor Fantasy League - OpenAPI 3.0.3
 * This is the Survivor Fantasy League Spec for API's
 *
 * OpenAPI spec version: 1.0.11
 * Contact: pklein111697@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface CreateSeasonRequestBody
 */
export interface CreateSeasonRequestBody {

    /**
     * The number of the season.
     *
     * @type {number}
     * @memberof CreateSeasonRequestBody
     * @example 1
     */
    seasonNumber: number;

    /**
     * Whether the season is active.
     *
     * @type {boolean}
     * @memberof CreateSeasonRequestBody
     * @example true
     */
    isActive: boolean;

    /**
     * The theme of the season.
     *
     * @type {string}
     * @memberof CreateSeasonRequestBody
     * @example Survivor: Borneo
     */
    theme: string;

    /**
     * The location of the season.
     *
     * @type {string}
     * @memberof CreateSeasonRequestBody
     * @example Borneo
     */
    location: string;

    /**
     * The name of the season.
     *
     * @type {string}
     * @memberof CreateSeasonRequestBody
     * @example Survivor: Borneo
     */
    name: string;

    /**
     * The number of contestants in the season.
     *
     * @type {number}
     * @memberof CreateSeasonRequestBody
     * @example 16
     */
    numberOfContestants: number;

    /**
     * The start date of the season.
     *
     * @type {string}
     * @memberof CreateSeasonRequestBody
     * @example Wed May 31 00:00:00 UTC 2000
     */
    startDate: string;

    /**
     * The end date of the season.
     *
     * @type {string}
     * @memberof CreateSeasonRequestBody
     * @example Wed Aug 23 00:00:00 UTC 2000
     */
    endDate: string;

    /**
     * The logo image of the season.
     *
     * @type {Blob}
     * @memberof CreateSeasonRequestBody
     */
    seasonLogo: Blob;
}